software.sh

+++ software/software.sh
@@ -16,7 +16,7 @@ export LFS_TGT=x86_64-pc-linux-gnu
 mkdir -p $HOME/bin
 cp newmod.sh $HOME/bin
 
-#rm -rf $MODS $HOME/.R_LIBS_USER $HOME/.Python
+#rm -rf $MODS #$HOME/.R_LIBS_USER $HOME/.Python
 #rm -rf $HOME/.R_LIBS_USER $HOME/.Python
 #rm -rf $HOME/.Python
 
@@ -35,10 +35,12 @@ module load autotools
 module load cmake/3.6
 module load gcc/6.2
 
-echo 'java-8.0.111'
-echo '#!/bin/bash
-module list
-cd $SOUR && wget -O l.tar.gz http://javadl.oracle.com/webapps/download/AutoDL?BundleId=216424 && \
+
+if [ ! -f $MODF/general/java/8.0.111 ]; then 
+	echo 'java-8.0.111'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && wget -O l.tar.gz http://javadl.oracle.com/webapps/download/AutoDL?BundleId=216424 && \
     mv l.tar.gz jre-8.0.111-linux-x64.tar.gz && \
     tar -zxvf jre-8.0.111-linux-x64.tar.gz && \
     cd jre1.8.0_111 && \
@@ -49,14 +51,16 @@ cd $SOUR && wget -O l.tar.gz http://javadl.oracle.com/webapps/download/AutoDL?Bu
     -p $MODF/general/ \
     -v 8.0.111 \
     -d 8.0.111
-' > $LOGS/java-8.0.111.sh
-chmod 755 $LOGS/java-8.0.111.sh
-srun -o $LOGS/java-8.0.111.out $LOGS/java-8.0.111.sh
+	' > $LOGS/java-8.0.111.sh
+	chmod 755 $LOGS/java-8.0.111.sh
+	srun -o $LOGS/java-8.0.111.out $LOGS/java-8.0.111.sh
+fi
 
-echo 'libevent-2.0.22'
-echo '#!/bin/bash
-module list
-cd $SOUR && wget -O l.tar.gz https://github.com/libevent/libevent/archive/release-2.0.22-stable.tar.gz && \
+if [ ! -f $MODF/libs/libevent/2.0.22 ]; then
+	echo 'libevent-2.0.22'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && wget -O l.tar.gz https://github.com/libevent/libevent/archive/release-2.0.22-stable.tar.gz && \
     mv l.tar.gz release-2.0.22-stable.tar.gz && \
     tar -zxvf release-2.0.22-stable.tar.gz && \
     cd libevent-release-2.0.22-stable && ./autogen.sh && \
@@ -67,17 +71,18 @@ cd $SOUR && wget -O l.tar.gz https://github.com/libevent/libevent/archive/releas
     -p $MODF/libs/ \
     -v 2.0.22 \
     -d 2.0.22
-' > $LOGS/libevent-2.0.22.sh
-chmod 755 $LOGS/libevent-2.0.22.sh
-srun -o $LOGS/libevent-2.0.22.out $LOGS/libevent-2.0.22.sh
-
+	' > $LOGS/libevent-2.0.22.sh
+	chmod 755 $LOGS/libevent-2.0.22.sh
+	srun -o $LOGS/libevent-2.0.22.out $LOGS/libevent-2.0.22.sh
+fi
 module load libevent/2.0.22
 
-echo 'ncurses-6.0'
-echo '#!/bin/bash
-module list
-rm -rf $SOUR/ncurses-6.0
-cd $SOUR && \
+if [ ! -f $MODF/libs/ncurses/6.0 ]; then
+	echo 'ncurses-6.0'
+	echo '#!/bin/bash
+	module list
+	rm -rf $SOUR/ncurses-6.0
+	cd $SOUR && \
     wget -O d.tgz http://ftp.gnu.org/pub/gnu/ncurses/ncurses-6.0.tar.gz && \
     mv d.tgz ncurses-6.0.tar.gz && \
     tar -zxvf ncurses-6.0.tar.gz && \
@@ -89,33 +94,35 @@ cd $SOUR && \
     -p $MODF/libs/ \
     -v 6.0 \
     -d 6.0
-' > $LOGS/ncurses-6.0.sh
-chmod 755 $LOGS/ncurses-6.0.sh
-srun -o $LOGS/ncurses-6.0.out $LOGS/ncurses-6.0.sh
+	' > $LOGS/ncurses-6.0.sh
+	chmod 755 $LOGS/ncurses-6.0.sh
+	srun -o $LOGS/ncurses-6.0.out $LOGS/ncurses-6.0.sh
+fi
 
 module load ncurses/6.0
 
-echo 'bzip2-1.0.6'
-echo '#!/bin/bash
-module list
-rm -rf $SOUR/bzip2-1.0.6 $SOFT/bzip2/1.0.6
-cd $SOUR && \
+if [ ! -f $MODF/libs/bzip2/1.0.6 ]; then
+	echo 'bzip2-1.0.6'
+	echo '#!/bin/bash
+	module list
+	rm -rf $SOUR/bzip2-1.0.6 $SOFT/bzip2/1.0.6
+	cd $SOUR && \
     wget -O d.tar.gz http://www.bzip.org/1.0.6/bzip2-1.0.6.tar.gz && \
     mv d.tar.gz bzip2-1.0.6.tar.gz && \
     tar -xzf bzip2-1.0.6.tar.gz && \
     cd bzip2-1.0.6 && \
     mkdir -p $SOFT/bzip2/1.0.6 && \
-    #sed -i "18s/.*/CC\=gcc -fPIC/" Makefile  && \
-    #COMPILE_FLAGS+=-fPIC make Makefile-libbz2_so && COMPILE_FLAGS+=-fPIC make clean && COMPILE_FLAGS+=-fPIC make && \
-    #COMPILE_FLAGS+=-fPIC make -n install PREFIX=$SOFT/bzip2/1.0.6 && \
-    #COMPILE_FLAGS+=-fPIC make install PREFIX=$SOFT/bzip2/1.0.6' > $LOGS/bzip2-1.0.6.sh
-echo "sed -i 's@\(ln -s -f \)\$(PREFIX)/bin/@\1@' Makefile" >> $LOGS/bzip2-1.0.6.sh
-echo 'sed -i "s@(PREFIX)/man@(PREFIX)/share/man@g" Makefile
-    make -f Makefile-libbz2_so
-    make clean
-    make
-    make PREFIX=$SOFT/bzip2/1.0.6 install 
-    cp -v bzip2-shared $SOFT/bzip2/1.0.6/bin/bzip2
+    sed -i "18s/.*/CC\=gcc -fPIC/" Makefile  && \
+    COMPILE_FLAGS+=-fPIC make Makefile-libbz2_so && COMPILE_FLAGS+=-fPIC make clean && COMPILE_FLAGS+=-fPIC make && \
+    COMPILE_FLAGS+=-fPIC make -n install PREFIX=$SOFT/bzip2/1.0.6 && \
+    COMPILE_FLAGS+=-fPIC make install PREFIX=$SOFT/bzip2/1.0.6' > $LOGS/bzip2-1.0.6.sh
+	#echo "sed -i 's@\(ln -s -f \)\$(PREFIX)/bin/@\1@' Makefile" >> $LOGS/bzip2-1.0.6.sh
+	#echo 'sed -i "s@(PREFIX)/man@(PREFIX)/share/man@g" Makefile
+	#echo 'make -f Makefile-libbz2_so
+ 	#   make clean
+ 	#   make
+ 	#   make PREFIX=$SOFT/bzip2/1.0.6 install 
+	echo 'cp -v bzip2-shared $SOFT/bzip2/1.0.6/bin/bzip2
     cp -av libbz2.so* $SOFT/bzip2/1.0.6/lib 
     newmod.sh \
     -s bzip2 \
@@ -123,15 +130,17 @@ echo 'sed -i "s@(PREFIX)/man@(PREFIX)/share/man@g" Makefile
     -v 1.0.6 \
     -d 1.0.6
     ' >> $LOGS/bzip2-1.0.6.sh
-chmod 755 $LOGS/bzip2-1.0.6.sh
-srun -o $LOGS/bzip2-1.0.6.out $LOGS/bzip2-1.0.6.sh #-o $LOGS/bzip2-1.0.6.out
-
+	chmod 755 $LOGS/bzip2-1.0.6.sh
+	srun -o $LOGS/bzip2-1.0.6.out $LOGS/bzip2-1.0.6.sh #-o $LOGS/bzip2-1.0.6.out
+fi
 module load bzip2/1.0.6
 
-echo 'xz-5.2.2'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+
+if [ ! -f $MODF/libs/xz/5.2.2 ]; then
+	echo 'xz-5.2.2'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz http://tukaani.org/xz/xz-5.2.2.tar.gz && \
     mv d.tar.gz xz-5.2.2.tar.gz && \
     tar -xzf xz-5.2.2.tar.gz && \
@@ -144,15 +153,16 @@ cd $SOUR && \
     -v 5.2.2 \
     -d 5.2.2
     ' > $LOGS/xz-5.2.2.sh
-chmod 755 $LOGS/xz-5.2.2.sh
-srun -o $LOGS/xz-5.2.2.out $LOGS/xz-5.2.2.sh
-
+	chmod 755 $LOGS/xz-5.2.2.sh
+	srun -o $LOGS/xz-5.2.2.out $LOGS/xz-5.2.2.sh
+fi
 module load xz/5.2.2
 
-echo 'pcre-8.39'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/libs/pcre/8.39 ]; then
+	echo 'pcre-8.39'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz ftp://ftp.csx.cam.ac.uk/pub/software/programming/pcre//pcre-8.39.tar.gz && \
     mv d.tar.gz pcre-8.39.tar.gz && \
     tar -xzf pcre-8.39.tar.gz && \
@@ -165,15 +175,16 @@ cd $SOUR && \
     -v 8.39 \
     -d 8.39
     ' > $LOGS/pcre-8.39.sh
-chmod 755 $LOGS/pcre-8.39.sh
-srun -o $LOGS/pcre-8.39.out $LOGS/pcre-8.39.sh # -o $LOGS/pcre-8.39.out
-
+	chmod 755 $LOGS/pcre-8.39.sh
+	srun -o $LOGS/pcre-8.39.out $LOGS/pcre-8.39.sh # -o $LOGS/pcre-8.39.out
+fi
 module load pcre/8.39
 
-echo 'curl-7.51.0'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/libs/curl/7.51.0 ]; then
+	echo 'curl-7.51.0'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz https://curl.haxx.se/download/curl-7.51.0.tar.gz && \
     mv d.tar.gz curl-7.51.0.tar.gz && \
     tar -xzf curl-7.51.0.tar.gz && \
@@ -186,15 +197,16 @@ cd $SOUR && \
     -v 7.51.0 \
     -d 7.51.0
     ' > $LOGS/curl-7.51.0.sh
-chmod 755 $LOGS/curl-7.51.0.sh
-srun -o $LOGS/curl-7.51.0.out $LOGS/curl-7.51.0.sh # -o $LOGS/pcre-8.39.out
-
+	chmod 755 $LOGS/curl-7.51.0.sh
+	srun -o $LOGS/curl-7.51.0.out $LOGS/curl-7.51.0.sh # -o $LOGS/pcre-8.39.out
+fi
 module load curl/7.51.0
 
-echo 'openblas-0.2.19'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/libs/openblas/0.2.19 ]; then
+	echo 'openblas-0.2.19'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz https://github.com/xianyi/OpenBLAS/archive/v0.2.19.tar.gz && \
     mv d.tar.gz openblas-0.2.19.tar.gz && \
     tar -xzf openblas-0.2.19.tar.gz && \
@@ -208,16 +220,17 @@ cd $SOUR && \
     -v 0.2.19 \
     -d 0.2.19
     ' > $LOGS/openblas-0.2.19.sh
-chmod 755 $LOGS/openblas-0.2.19.sh
-srun -o $LOGS/openblas-0.2.19.out $LOGS/openblas-0.2.19.sh # -o $LOGS/openblas-0.2.19.out
-
+	chmod 755 $LOGS/openblas-0.2.19.sh
+	srun -o $LOGS/openblas-0.2.19.out $LOGS/openblas-0.2.19.sh # -o $LOGS/openblas-0.2.19.out
+fi
 module load openblas/0.2.19
 
-echo 'rlang-3.3.2'
-echo '#!/bin/bash
-module list
-rm -rf $SOUR/R-3.3.2 $SOFT/r/3.3.2
-cd $SOUR && \
+if [ ! -f $MODF/general/rlang/3.3.2 ]; then
+	echo 'rlang-3.3.2'
+	echo '#!/bin/bash
+	module list
+	rm -rf $SOUR/R-3.3.2 $SOFT/r/3.3.2
+	cd $SOUR && \
     wget -O d.tar.gz http://ftp5.gwdg.de/pub/misc/cran/src/base/R-3/R-3.3.2.tar.gz && \
     mv d.tar.gz R-3.3.2.tar.gz && \
     tar -xzf R-3.3.2.tar.gz && \
@@ -240,14 +253,16 @@ cd $SOUR && \
     echo "prepend-path CPLUS_INCLUDE_PATH $SOFT/r/3.3.2/lib64/R/include" >> $MODF/general/rlang/3.3.2
     echo "prepend-path OBJC_INCLUDE_PATH $SOFT/r/3.3.2/lib64/R/include" >> $MODF/general/rlang/3.3.2
     ' > $LOGS/rlang-3.3.2.sh
-chmod 755 $LOGS/rlang-3.3.2.sh
-srun -o $LOGS/rlang-3.3.2.out $LOGS/rlang-3.3.2.sh #-o $LOGS/r-3.3.2.out
+	chmod 755 $LOGS/rlang-3.3.2.sh
+	srun -o $LOGS/rlang-3.3.2.out $LOGS/rlang-3.3.2.sh #-o $LOGS/r-3.3.2.out
+fi
 
-echo 'python-2.7.12'
-echo '#!/bin/bash
-module list
-rm -rf $SOUR/python-2.7.12 $SOFT/python/2.7.12
-cd $SOUR && \
+if [ ! -f $MODF/general/python/2.7.12 ]; then
+	echo 'python-2.7.12'
+	echo '#!/bin/bash
+	module list
+	rm -rf $SOUR/python-2.7.12 $SOFT/python/2.7.12
+	cd $SOUR && \
     wget -O d.tgz https://www.python.org/ftp/python/2.7.12/Python-2.7.12.tgz && \
     mv d.tgz python-2.7.12.tgz && \
     tar xzf python-2.7.12.tgz && \
@@ -274,13 +289,15 @@ cd $SOUR && \
     echo "setenv PYTHONHOME $SOFT/python/2.7.12/" >> $MODF/general/python/2.7.12
     echo "setenv PYTHONUSERBASE \$home/.Python/2.7.12/" >> $MODF/general/python/2.7.12
     ' > $LOGS/python-2.7.12.sh
-chmod 755 $LOGS/python-2.7.12.sh
-srun -o $LOGS/python-2.7.12.out $LOGS/python-2.7.12.sh #-o $LOGS/python-2.7.12.out
+	chmod 755 $LOGS/python-2.7.12.sh
+	srun -o $LOGS/python-2.7.12.out $LOGS/python-2.7.12.sh #-o $LOGS/python-2.7.12.out
+fi
 
-echo 'pigz-2.3.4'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/general/pigz/2.3.4 ]; then
+	echo 'pigz-2.3.4'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget http://zlib.net/pigz/pigz-2.3.4.tar.gz && \
     tar -zxvf pigz-2.3.4.tar.gz && \
     cd pigz-2.3.4 && make && \
@@ -292,14 +309,16 @@ cd $SOUR && \
     -p $MODF/general/ \
     -v 2.3.4 \
     -d 2.3.4
-' > $LOGS/pigz-2.3.4.sh
-chmod 755 $LOGS/pigz-2.3.4.sh
-srun -o $LOGS/pigz-2.3.4.out $LOGS/pigz-2.3.4.sh
+	' > $LOGS/pigz-2.3.4.sh
+	chmod 755 $LOGS/pigz-2.3.4.sh
+	srun -o $LOGS/pigz-2.3.4.out $LOGS/pigz-2.3.4.sh
+fi
 
-echo 'tmux-2.3'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/general/tmux/2.3 ]; then
+	echo 'tmux-2.3'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O t.tar.gz https://github.com/tmux/tmux/releases/download/2.3/tmux-2.3.tar.gz && \
     mv t.tar.gz tmux-2.3.tar.gz && \
     tar -zxvf tmux-2.3.tar.gz && \
@@ -314,15 +333,16 @@ cd $SOUR && \
     -d 2.3
     echo "module load ncurses/6.0" >> $MODF/general/tmux/2.3
     echo "module load libevent/2.0.22" >> $MODF/general/tmux/2.3
-' > $LOGS/tmux-2.3.sh
-chmod 755 $LOGS/tmux-2.3.sh
-srun -o $LOGS/tmux-2.3.out $LOGS/tmux-2.3.sh
+	' > $LOGS/tmux-2.3.sh
+	chmod 755 $LOGS/tmux-2.3.sh
+	srun -o $LOGS/tmux-2.3.out $LOGS/tmux-2.3.sh
+fi
 
-
-echo 'bowtie2-2.2.9'
-echo '#!/bin/bash
-module list
-cd $SOUR
+if [ ! -f $MODF/bioinformatics/bowtie/2.2.9 ]; then
+	echo 'bowtie2-2.2.9'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR
     wget -O b.zip https://sourceforge.net/projects/bowtie-bio/files/bowtie2/2.2.9/bowtie2-2.2.9-linux-x86_64.zip
     mv b.zip bowtie2-2.2.9-linux-x86_64.zip
     rm -rf bowtie2-2.2.9
@@ -336,14 +356,16 @@ cd $SOUR
     -v 2.2.9 \
     -d 2.2.9
     #echo "module load libevent/2.0.22" >> $MODF/bioinformatics/bowtie/2.2.9
-' > $LOGS/bowtie-2.2.9.sh
-chmod 755 $LOGS/bowtie-2.2.9.sh
-srun -o $LOGS/bowtie-2.2.9.out $LOGS/bowtie-2.2.9.sh
+	' > $LOGS/bowtie-2.2.9.sh
+	chmod 755 $LOGS/bowtie-2.2.9.sh
+	srun -o $LOGS/bowtie-2.2.9.out $LOGS/bowtie-2.2.9.sh
+fi
 
-echo 'cufflinks-2.2.1'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/bioinformatics/cufflinks/2.2.1 ]; then
+	echo 'cufflinks-2.2.1'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz http://cole-trapnell-lab.github.io/cufflinks/assets/downloads/cufflinks-2.2.1.Linux_x86_64.tar.gz && \
     mv d.tar.gz cufflinks-2.2.1.Linux_x86_64.tar.gz && \
     tar -zxvf cufflinks-2.2.1.Linux_x86_64.tar.gz && \
@@ -355,14 +377,16 @@ cd $SOUR && \
     -p $MODF/bioinformatics/ \
     -v 2.2.1 \
     -d 2.2.1
-' > $LOGS/cufflinks-2.2.1.sh
-chmod 755 $LOGS/cufflinks-2.2.1.sh
-srun -o $LOGS/cufflinks-2.2.1.out $LOGS/cufflinks-2.2.1.sh
+	' > $LOGS/cufflinks-2.2.1.sh
+	chmod 755 $LOGS/cufflinks-2.2.1.sh
+	srun -o $LOGS/cufflinks-2.2.1.out $LOGS/cufflinks-2.2.1.sh
+fi
 
-echo 'fastqc-0.11.5'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/bioinformatics/fastqc/0.11.5 ]; then
+	echo 'fastqc-0.11.5'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.zip http://www.bioinformatics.babraham.ac.uk/projects/fastqc/fastqc_v0.11.5.zip && \
     mv d.zip fastqc_v0.11.5.zip && \
     unzip fastqc_v0.11.5.zip && \
@@ -376,37 +400,17 @@ cd $SOUR && \
     -v 0.11.5 \
     -d 0.11.5
     echo "module load java" >> $MODF/bioinformatics/fastqc/0.11.5
-' > $LOGS/fastqc-0.11.5.sh
-chmod 755 $LOGS/fastqc-0.11.5.sh
-srun -o $LOGS/fastqc-0.11.5.out $LOGS/fastqc-0.11.5.sh
+	' > $LOGS/fastqc-0.11.5.sh
+	chmod 755 $LOGS/fastqc-0.11.5.sh
+	srun -o $LOGS/fastqc-0.11.5.out $LOGS/fastqc-0.11.5.sh
+fi
 
-echo 'flexbar-2.5.0'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
-    wget -O d.tgz https://github.com/seqan/flexbar/releases/download/v2.5.0/flexbar_v2.5_linux64.tgz && \
-    mv d.tgz flexbar_v2.5_linux64.tgz && \
-    tar zxvf flexbar_v2.5_linux64.tgz && \
-    cd flexbar_v2.5_linux64/ && \
-    mkdir -p $SOFT/flexbar/2.5.0/bin && \
-    cp -r * $SOFT/flexbar/2.5.0/bin && \
-    mkdir -p $SOFT/flexbar/2.5.0/lib && \
-    cp -r libtbb.so.2 $SOFT/flexbar/2.5.0/lib && \
-    newmod.sh \
-    -s flexbar \
-    -p $MODF/bioinformatics/ \
-    -v 2.5.0 \
-    -d 2.5.0
-    echo "module load bzip2" >> $MODF/bioinformatics/flexbar/2.5.0
-' > $LOGS/flexbar-2.5.0.sh
-chmod 755 $LOGS/flexbar-2.5.0.sh
-srun -o $LOGS/flexbar-2.5.0.out $LOGS/flexbar-2.5.0.sh
-
-echo 'samtools-1.3.1'
-echo '#!/bin/bash
-module list
-rm -rf $SOUR/samtools-1.3.1
-cd $SOUR && \
+if [ ! -f $MODF/bioinformatics/samtools/1.3.1 ]; then
+	echo 'samtools-1.3.1'
+	echo '#!/bin/bash
+	module list
+	rm -rf $SOUR/samtools-1.3.1
+	cd $SOUR && \
     wget -O d.tar.bz2 https://github.com/samtools/samtools/releases/download/1.3.1/samtools-1.3.1.tar.bz2 && \
     mv d.tar.bz2 samtools-1.3.1.tar.bz2 && \
     tar -jxvf samtools-1.3.1.tar.bz2 && \
@@ -418,14 +422,16 @@ cd $SOUR && \
     -p $MODF/bioinformatics/ \
     -v 1.3.1 \
     -d 1.3.1
-' > $LOGS/samtools-1.3.1.sh
-chmod 755 $LOGS/samtools-1.3.1.sh
-srun -o $LOGS/samtools-1.3.1.out $LOGS/samtools-1.3.1.sh
+	' > $LOGS/samtools-1.3.1.sh
+	chmod 755 $LOGS/samtools-1.3.1.sh
+	srun -o $LOGS/samtools-1.3.1.out $LOGS/samtools-1.3.1.sh
+fi
 
-echo 'hisat-2.0.4'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/bioinformatics/hisat/2.0.4 ]; then
+	echo 'hisat-2.0.4'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.zip ftp://ftp.ccb.jhu.edu/pub/infphilo/hisat2/downloads/hisat2-2.0.4-Linux_x86_64.zip && \
     mv d.zip hisat2-2.0.4-Linux_x86_64.zip && \
     unzip hisat2-2.0.4-Linux_x86_64.zip && \
@@ -437,14 +443,16 @@ cd $SOUR && \
     -p $MODF/bioinformatics/ \
     -v 2.0.4 \
     -d 2.0.4
-' > $LOGS/hisat-2.0.4.sh
-chmod 755 $LOGS/hisat-2.0.4.sh
-srun -o $LOGS/hisat-2.0.4.out $LOGS/hisat-2.0.4.sh
+	' > $LOGS/hisat-2.0.4.sh
+	chmod 755 $LOGS/hisat-2.0.4.sh
+	srun -o $LOGS/hisat-2.0.4.out $LOGS/hisat-2.0.4.sh
+fi
 
-echo 'stringtie-1.3.0'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/bioinformatics/stringtie/1.3.0 ]; then
+	echo 'stringtie-1.3.0'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz http://ccb.jhu.edu/software/stringtie/dl/stringtie-1.3.0.Linux_x86_64.tar.gz && \
     mv d.tar.gz stringtie-1.3.0.Linux_x86_64.tar.gz && \
     tar -zxvf stringtie-1.3.0.Linux_x86_64.tar.gz && \
@@ -456,14 +464,16 @@ cd $SOUR && \
     -p $MODF/bioinformatics/ \
     -v 1.3.0 \
     -d 1.3.0
-' > $LOGS/stringtie-1.3.0.sh
-chmod 755 $LOGS/stringtie-1.3.0.sh
-srun -o $LOGS/stringtie-1.3.0.out $LOGS/stringtie-1.3.0.sh
+	' > $LOGS/stringtie-1.3.0.sh
+	chmod 755 $LOGS/stringtie-1.3.0.sh
+	srun -o $LOGS/stringtie-1.3.0.out $LOGS/stringtie-1.3.0.sh
+fi
 
-echo 'bwa-0.7.15'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/bioinformatics/bwa/0.7.15 ]; then
+	echo 'bwa-0.7.15'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz https://github.com/lh3/bwa/archive/v0.7.15.tar.gz && \
     mv d.tar.gz v0.7.15.tar.gz && \
     tar -zxvf v0.7.15.tar.gz && \
@@ -476,14 +486,16 @@ cd $SOUR && \
     -p $MODF/bioinformatics/ \
     -v 0.7.15 \
     -d 0.7.15
-' > $LOGS/bwa-0.7.15.sh
-chmod 755 $LOGS/bwa-0.7.15.sh
-srun -o $LOGS/bwa-0.7.15.out $LOGS/bwa-0.7.15.sh
+	' > $LOGS/bwa-0.7.15.sh
+	chmod 755 $LOGS/bwa-0.7.15.sh
+	srun -o $LOGS/bwa-0.7.15.out $LOGS/bwa-0.7.15.sh
+fi
 
-echo 'tophat-2.1.1'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/bioinformatics/tophat/2.1.1 ]; then
+	echo 'tophat-2.1.1'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz https://ccb.jhu.edu/software/tophat/downloads/tophat-2.1.1.Linux_x86_64.tar.gz && \
     mv d.tar.gz tophat-2.1.1.Linux_x86_64.tar.gz && \
     tar -zxvf tophat-2.1.1.Linux_x86_64.tar.gz && \
@@ -495,14 +507,16 @@ cd $SOUR && \
     -p $MODF/bioinformatics/ \
     -v 2.1.1 \
     -d 2.1.1
-' > $LOGS/tophat-2.1.1.sh
-chmod 755 $LOGS/tophat-2.1.1.sh
-srun -o $LOGS/tophat-2.1.1.out $LOGS/tophat-2.1.1.sh
+	' > $LOGS/tophat-2.1.1.sh
+	chmod 755 $LOGS/tophat-2.1.1.sh
+	srun -o $LOGS/tophat-2.1.1.out $LOGS/tophat-2.1.1.sh
+fi
 
-echo 'star-2.5.2b'
-echo '#!/bin/bash
-module list
-cd $SOUR && \
+if [ ! -f $MODF/bioinformatics/star/2.5.2b ]; then
+	echo 'star-2.5.2b'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \
     wget -O d.tar.gz https://github.com/alexdobin/STAR/archive/2.5.2b.tar.gz && \
     mv d.tar.gz 2.5.2b.tar.gz && \
     tar -xzf 2.5.2b.tar.gz && \
@@ -514,14 +528,16 @@ cd $SOUR && \
     -p $MODF/bioinformatics/ \
     -v 2.5.2b \
     -d 2.5.2b
-' > $LOGS/star-2.5.2b.sh
-chmod 755 $LOGS/star-2.5.2b.sh
-srun -o $LOGS/star-2.5.2b.out $LOGS/star-2.5.2b.sh
+	' > $LOGS/star-2.5.2b.sh
+	chmod 755 $LOGS/star-2.5.2b.sh
+	srun -o $LOGS/star-2.5.2b.out $LOGS/star-2.5.2b.sh
+fi
 
-echo 'skewer-0.2.2'
-echo '#!/bin/bash
-module list
-cd $SOUR && \ 
+if [ ! -f $MODF/bioinformatics/skewer/0.2.2 ]; then
+	echo 'skewer-0.2.2'
+	echo '#!/bin/bash
+	module list
+	cd $SOUR && \ 
     wget http://downloads.sourceforge.net/project/skewer/Binaries/skewer-0.2.2-linux-x86_64
     mkdir -p $SOFT/skewer/0.2.2/bin
     cp skewer-0.2.2-linux-x86_64 $SOFT/skewer/0.2.2/bin/skewer
@@ -531,10 +547,10 @@ cd $SOUR && \
     -p $MODF/bioinformatics/ \
     -v 0.2.2 \
     -d 0.2.2
-' > $LOGS/skewer-0.2.2.sh
-chmod 755 $LOGS/skewer-0.2.2.sh
-srun -o $LOGS/skewer-0.2.2.out $LOGS/skewer-0.2.2.sh
-
+	' > $LOGS/skewer-0.2.2.sh
+	chmod 755 $LOGS/skewer-0.2.2.sh
+	srun -o $LOGS/skewer-0.2.2.out $LOGS/skewer-0.2.2.sh
+fi
 
 chmod -R 755 $MODS
